Implemented support for CP/M ROM in SVI-707.
The SVI-707 extension now also needs the 4kB CP/M (boot) ROM to work properly.
To support this cleanly, I added an optional parameter to MSXFDC's constructor
where you can specify the rom ID to be used for the disk ROM. By default this
is still the empty string as it was before.

Leonard Oliveira confirmed that the initial state of the cpmRomEnabled flipflop
is indeed such that the CP/M ROM is enabled, which thus also acts as a boot
ROM. He also confirmed that the reset signal will set it back to this initial
value. This behaviour is also part of this implementation.