Change HD image parsing/opening
Two changes:

1) Before this patch when the -hda commandline option was specified, we first
(tried to) open the default image 'hd.dsk' and only later switch to the
specified image. Now we immediately open the specified image.

2) Before this patch the 'HD' constructor tried to open the default image
'hd.dsk', but if that image doesn't exist yet, creating it is postponed till
it's actually required.

Part of the reason to postponed in 1) is that when a -hda command line image is
specified we don't want to create the (unused) default image. With the above
two changes this is no longer a problem.

Another reason mentioned in the comments in the code is to avoid creating the
default image in case of '-testconfig'. Though I'm not sure that's a valid
reason: in case the default image cannot be created (e.g. disk full) then the
ide extension should be marked as 'not working'.

Though the main motivation for this patch is to fix the crash that will be
described in more detail in the next (follow-up) patch.