Fix #1011 Uncaught exception if no hard disk image can be created
The previous patch would always open or create a harddisk image in the 'HD'
constructor. Or throw an exception when the specified (command line option)
image could not be found or when the default image could not be created.

This fixed bug #1011. Because when we instead lazily create the default image
and something goes wrong then (e.g. on query the emulated disk size) we've
already started emulation and that code path is not prepared to handle these
errors.

The previous patch immediately opened/created the image. This patch removes the
(no longer needed) code that lazily created the image.