Fix possible crash on reset when there was still a VDP cmd executing
The VDPCmdEngine::reset() method does among other things the following:
 - Set 'scrMode' = -1. This indicates a screen mode in which no command can be
   executed (e.g. text mode on V9938).
 - Reset registers and abort the current command by writing 0 to all command
   registers. Writing 0 to R#46 is the ABORT command.

This is fine except for the order in which these things happen. Writing the
command registers is done from low to high (R#32 -> R#46). So the command is
only aborted on the last write. Thus the earlier writes are processed with a
command that's possibly still in progress. That's all fine because no
(emulated) time is passing during this sequence. However the assignment
scrMode=-1 has also already happened. So during processing the early register
writes the invariant
  scrMode == -1   ==>   no command is executing
can be temporarily violated. And this can trigger an assert in some (rare)
circumstances. In a non-assert build everything would be fine.

Once the problem is understood it's easy to fix by changing the order of the
items mentioned above. Basically one of the first things we do now is to
actually abort the command.