Fix bug #1026 Reverse removes cassette, when using -cassetteplayer option
See the bug report for a description of the scenario:
  https://github.com/openMSX/openMSX/issues/1026

This is what's going on internally:
 1) OpenMSX starts, the msx machine and CassettePlayer objects are created
 2) The command line is parsed, this inserts the tape (via a RecordedCommand)
 3) Reverse is auto-started via a Tcl script
 4) The CassettePlayer is auto-plugged via a Tcl script

* When 2 is executed reverse is not yet enabled, so the insert-tape command is
  not yet recorded. This shouldn't matter because an initial snapshot is
  created at the moment reverse does get enabled.
* When 3 is executed the CassettePlayer is not yet plugged in. (Before this
  patch) this means that the state of the CassettePlayer is not included in the
  initial snapshot.
* When 4 is executed reverse is already enabled, thus the plug command does get
  recorded in the replay information.
So when we reverse all the way to the initial snapshot, we will re-plug the
CassettePlayer, but it won't have any information about the cassette image that
was inserted.

As is so often the case, once you understand the problem, fixing it isn't that
hard. So this patch will serialize the state of the CassettePlayer also if it's
not currently plugged in.

But we have more pluggables. Shouldn't we do this for more than only the
CassettePlayer? It's been a long time since I wrote this code, but I guess at
the time I was mostly thinking about mouse and Joystick, and those devices
don't retain any state if they are not plugged in (because e.g. the mouse is
not powered). In a cassette player you physically insert a tape, and this is of
course 'remembered' even if the cassette player is not powered on.