Fix #1027 TAB completion of help command shouldn't list commands with :: prefix
Disable tab-completion for explicitly global namespace qualified commands. This
was intentionally added as (part of) commit 39fb861, and this is a valid way to
enter Tcl commands. But I do agree it's not likely a user will ever enter
commands like this. And it does clutter the output for the scenarion described
in the bug report.