Replaced some '(unsigned) long' types with '(s)size_t'
The intention of this code was to use a type with the same size as a pointer.
Not for correctness but because gcc generated slightly better code for it (less
zero-extension instructions). Using the type 'long' for this works perfectly on
Linux (it's 32/64 bits on respectively 32/64-bit systems). But on windows
'long' is always 32-bit (also on 64-bit systems). However 'size_t' does work.

Suggested by clang-tidy [google-runtime-int].