Fixed autoruncassettes feature.
Due to the changes I made in the type command in the previous commits, the
autoruncassettes feature was broken. This was unexpected, because the type
command is still there, but now it's implemented in Tcl.

However, at the moment the autorun feature is active and wants to type the
loading instruction, the Tcl scripts are not loaded yet and so one of the type
commands (the last one) will give an error, as it's executed immediately.

As it seems to be cleaner to use keyboard typing in this case, let's just make
it explicitly use that method.