Added Toshiba HX-23F and HX-23 configuration thanks to FRS
FRS submitted these. He got a real HX-23F and created its config by looking at
the PCB and running test programs.

The HX-23 is very similar to the HX-23F. We assume here that the ROMs are
identical, which is also hinted by the blueMSX configuration of that machine.