Fix Moonsound sound effects in 'Life on Mars'
The final version of 'Life on Mars' contains a work around for this emulation
problem. We received a development version without this workaround. To reproduce
the bug in that version:
 - start 'sound bug.rom' with -ext moonsound
 - in the game select 'bomb tournament'
 - select the 3rd background music
Initially the sound-fx's are working, but they stop after a while.

Probably the sound-fx bug in 'Run Monster' is also fixed by this patch, but I
couldn't verify that yet. EDIT: I received my copy today, and I can confirm the
bug is indeed fixed.

See comments in code for a short description of the problem.

Based on my current (very slowly progressing) YM2413 reverse engineering work,
I'm guessing that the conditions in the real YMF278 hardware are simpler than
how it's currently emulated (again see comments in code). We should probably
revisit this in the future. But I think I should first finish the YM2413 work
before extending it to other sound chips.