Made output of 'help' command more compact
The help command prints a few lines followed by a list of approximately 75
command names. This long list ensures that the top lines will scroll off
screen. This patch formats that list in columns (like the output of
tab-completion), to increase the chance that the top lines remain visible. In
addition the list of commands is now alphabetically sorted.

Of course it's possible to scroll-up in the console by using PgUp. And the
console intro text even mentions this. But still 'help' is a command that will
mostly be used by beginners and they might not yet know about PgUp.

We could extend more commands with this group-in-columns feature. Though I
strongly suggest to be very selective in which commands we do extend, because
formatted output makes it very hard to use that output further in Tcl scripts.
E.g. we could format the output of 'machine_info pluggable', but that would
make it impossible to use this command in a for loop like this:
   foreach p [machine_info pluggable] { ...
On the other hand for the 'help' command there is no such problem:
- The extra lines at the top already make it hard to use the 'help' output in a
  script.
- There are better ways to query the list of available commands in a script
  (e.g. with 'info commands').