[1/12] Initial version of DeltaBlock based reverse snapshots
First the motivation for this change:
* We recently added the 'KonamiUltimateCollection' mapper. This contains a 8MB
  flashrom. When reverse is enabled, each second, we compress this 8MB memory
  block. In some cases this causes noticeable hiccups.
* On 'small' devices (e.g. Android) we don't enable reverse by default
  because it takes a relatively large amount of memory and CPU time.

In case of 'KonamiUltimateCollection' the flashrom rarely changes, so for each
reverse snapshot we spend time compressing the exact same memory block. This is
true in general, flashroms rarely change. But also for many other, sometimes
large, memory blocks like MoonSound sample RAM (typically gets (re-)loaded
infrequently) or Gfx9000 VRAM in case you're emulating non-v9990 software. Our
boosted machine configs do include both MoonSound and Gfx9000.

A smarter way to store memory blocks in reverse snapshots could be to only
store the difference between successive snapshots. In case the blocks truly
don't change much, this difference can be encoded very compact in memory,
resulting in lower overall memory consumption for the reverse system.
Calculating this difference is also much simpler compared to compressing a
block. So there's hope this can also be done faster.

There's one complication: if each snapshot stores the difference with the
previous snapshot, and you want to restore a specific snapshot, you have to
start from the initial snapshot and apply all differences in turn. This could
take a long time. Also to save memory we want to prune snapshots, that is from
recent history we keep more snapshots in memory compared to the more distant
history. But if all difference are linked in a chain it's impossible to drop
some of the snapshots.

To solve both problems we don't calculate the difference between two successive
snapshots but between a semi-fixed reference block and the current block. Once
this difference becomes too large we switch to a new reference.

This is only the first patch in a series. It adds the basic delta-compression
functionality. It's already fully working but:
 * Memory consumption isn't very good yet (although already better than before
   this patch. Of course very much depending on which emulated machine and
   software you use to measure).
 * Performance is actually a bit worse compared to before this patch.
Both issues will be addressed later in this patch series.

It's probably worth mentioning that this only affects the in-memory savestates
(the reverse snapshots). The on-disk savetstate and replay format is unchanged.