[2/12] Compress non-reference delta blocks
The previous patch explained the overall idea of delta-compression for reverse
snapshots. Let's go in a little more detail:

* When asked to include a memory block in a snapshot (an in-memory savestate)
  we first check whether we already have a reference for this block.
* If not we create a 'DeltaBlockCopy' object. This is basically a memcpy of the
  data.
* If we do have a reference we calculate the difference between that
  and the current block and store the result in a 'DeltaBlockDiff' object.
* (Typically) over time a memory block will start to deviate more and more from
  the reference, thus the delta between both becomes larger. When this
  difference becomes too large (using some heuristics to determine what 'too
  large' means exactly), instead of creating a new 'DeltaBlockDiff' we create a
  'DeltaBlockCopy' and use that as the new reference for the future.

Once we switch to a new reference block, the old reference isn't used anymore.
Not even when going back in time. In other words: we do use old reference
blocks to restore the machine state, but never as a reference to create new
delta-blocks). Instead, after a 'reverse goto' we always create new references.
We might loose a bit of compression efficiency because of this, but in return
we get a simpler implementation. (The underlying technical reason is that
blocks are identified based on their start address, and switching to a new
machine (internally done by 'reverse goto') creates new objects with different
start addresses for the memory blocks).

To be able to (efficiently) calculate the difference between memory blocks,
both blocks need to be present (uncompressed) in memory. Though after we switch
to a new reference we can compress the old reference block (see previous
paragraph). This gives the guarantee that the new reverse mechanism (delta
compression) never uses more memory compared to the old mechanism (compress
each block independently). (Modulo the size of a single snapshot and taking the
current heuristic into account. I'm not going into more detail).

I did some quick tests and the new mechanism indeed typically uses 2x-5x less
memory compared to the old one. As expected the gain is larger for more complex
MSX machines and/or when the MSX software doesn't use all included MSX devices.