[7/12] Optimize V9990VRAM serialize if unchanged
Gfx9000 has a relatively large amount of memory (512kB) and it is included in
our boosted machines. But MSX software for V9990 is relatively rare, so the
V9990VRAM content often doesn't change at all. This makes it a good candidate
for dirty tracking and null-serialization if not-dirty.

But even in case the user is running V9990 software with frequent VRAM writes
(of course V9990 commands also count as writes), the added overhead of
V9990VRAM dirty tracking should be negligible compared to the code surrounding
the write.