[3/12] Optimize delta calculation
A sub goal of this patch series is to improve the speed of the reverse system.
Before this patch this series actually made reverse slower (2%-50%(!), of
course very much depending on how/what you measure). With this patch I measured
a 4%-10% speedup. To be honest the speedups are not only the result of this
patch but also of some future patches in this series, but this patch had the
largest impact.

The core of the delta compression is a loop like this

    while ((q != end) && (*p == *q)) { ++p; ++q; }

It searches in 2 byte-buffers for the first difference or until the end of the
buffer is reached. There is a similar loop that searches for equal bytes. These
loops can be sped up in two ways:
 - Use sentinels to avoid the end-of-buffer check on each iteration.
 - Compare words- instead of bytes-at-a-time.
See comments in the code for much more details.

I measured that the optimized delta compression is indeed faster than the old
isolated blocks compression. But let's look at both from a distance:

Suppose we need to add a large block to the snapshot.
 * When compressing isolated blocks, the compression routine reads the whole
   input and produces some output. Some parts of the input might be read
   multiple times, but let's assume the compression routine is well optimized
   for memory caching and those secondary reads all come from L1 cache.
 * When using delta compression we compare two buffers and produce some output.
   Every byte of each input buffer is read exactly once. But still this routine
   needs to read twice as much data.
Depending on how compressible the input data is and how much it changed
compared to the previous snapshot, the size of the produced output can greatly
vary in size. But for simplicity let's ignore the output part, because
typically it's anyway an order of magnitude smaller than the input.

Very often memory bandwidth is the most important factor when optimizing code,
and from that perspective compressing isolated blocks seems better: it only
touches half the amount of memory. On the other hand delta compression has a
very simple memory access pattern (linear scan) while regular compression also
does some backwards memory reads (but only in a limited window).

When comparing both compression routines, delta- looks vastly simpler than
regular compression (we currently use 'snappy' as out compression algorithm,
but algorithms like 'LZ4' behave very similar). But still it only runs a little
faster (and the unoptimized version actually ran slower). So possibly this only
modest improvement can be explained by cache effects. There also is still room
to improve delta compression further, while snappy has reached its limits (or
is close to it).