[5/12] added TrackedRam wrapper around Ram
The previous patch added infrastructure to efficiently store a null-difference
in snapshots in case a memory block hasn't changed. This patch adds a helper
class that can actually detect when a Ram block hasn't changed.

This patch adds a 'TrackedRam' class. It offers almost the same interface as
the existing 'Ram' class, but it has an internal dirty-flag that keeps track of
whether there were any writes since the last reverse snapshot. So compared to
'Ram' it has a little overhead on each write operation.

IMHO an important property of this class is that when you change some existing
code from using 'Ram' to 'TrackedRam' it's impossible (or at least very hard)
to accidentally change the content of the Ram without also setting the internal
dirty flag. In case you do make a mistake (e.g. you miss some (indirect) write
when changing from Ram to TrackedRam) you get a compile error instead of wrong
runtime behavior.

Later patches in this series will actually start using this helper class.