[4/12] extend serialize_blob() with no-changes flag
The previous patch explained how delta-compression loops over 2 input buffers.
In case of the 'KonamiUltimateCollection' mapper this means looping over 2 8MB
buffers only to conclude that both buffers are still identical (in the vast
majority of the cases). From a memory cache perspective this is not a good
idea. And, also explained in the previous patch, memory access is often very
important for good performance.

If we have an alternative mechanism that can tell us the memory block hasn't
changed since the last snapshot we can avoid the memory scan. And then we can
directly store a null-difference in the snapshot. (Actually we don't store a
null-difference, instead we simply repeat the previously stored DeltaBlock).

The details of how to detect that blocks haven't changed are left for later
patches in this series. Instead this patch focusses on:
- How to actually store null-differences (e.g. we don't only need to keep track
  of a reference delta-block but also of the last delta-block). We also need to
  be able to indicate to the serializer that a block hasn't changed.
- What does the 'previous snapshot' mean exactly. In-memory savestates are made
  for more than one reason (e.g. loading a replay-file and actual reverse
  snapshots). Only the latter should be taken into account when defining the
  'previous snapshot'.