[6/12] Optimize SRAM serialization if not changed
The 'KonamiUltimateCollection' example was already used a few times in this
patch series. Very often the flashrom content doesn't change between reverse
snapshots. And we don't want to spend much time (and memory) storing this
(often unchanged) flashrom in the snapshot.

The previous few patches already made some preparations. This patch actually
optimizes reverse snapshots for flashroms in case they didn't change.

Basically the only non-trivial thing this patch does is to replace indirect
writes (write via a pointer) with explicit write() calls.

Actually as an internal implementation detail our 'AmdFlash' class uses a
'SRAM' object for the actual storage. In general most SRAMs are not written
that frequently, so they all benefit from the more efficient serialization when
not dirty. And even when they are written frequently (e.g. in some specific
period), dirty-tracking is not that expensive, so it also doesn't hurt.