[2/2] Permute rows in table
I've written a large comment block in the code, I won't repeat that here. But
very short summary:
  This patch tries to reorder the rows in the resample-hq coefficient table so
  that rows that are needed after each other (in time) are stored sequentially
  in memory.

Like with the previous patch the effect of this patch depends both on the host
system and the emulated configuration. On a i7 system with 8MB L3 cache I
couldn't measure much difference, but on a i5 with 4MB L3 cache I got an
additional 3% speedup (on top of the 4% of last patch).