[1a/2] Shift sinc functions over half an index
The 'hq'-resampler uses a large 2D table with pre-calculated (windowed) sinc
functions. This table has 4096 rows and in each row the sinc is shifted over
1/4096th of a 'step' (a step roughly corresponds to the duration of 1 host
sample (e.g. 44100Hz), see source code for more details).

The sinc function is defined as:
  sinc(x) := sin(pi * x) / (pi * x)
This is an even function (sinc(x) == sinc(-x)), So in principle we only need to
store one half of the function in the table.

This patch shifts the whole table over 1/8192th of a step. The reason for this
is to better exploit the symmetry that's present in the sinc function. I'll
explain by using a sine wave instead of a sinc, but the mechanism is the same
for both.

Suppose we want to store 8 samples of half a sine wave. One way to do this is
define:
   f1(n) := sin(n * pi/8)
Evaluating for n = 0..7 results in:
  n     :  0      1      2      3      4      5      6      7
  f1(n) : 0.000  0.383  0.707  0.924  1.000  0.924  0.707  0.383

Another possibility is:
   f2(n) := sin(n * pi/8 + pi/16)
Which results in:
  n     :  0      1      2      3      4      5      6      7
  f2(n) : 0.195  0.556  0.831  0.981  0.981  0.831  0.556  0.195

The table for f1() has 5 distinct value, for f2() there are only 4. Also the
formula to calculate the 2nd half of the table from the 1st half is slightly
simpler for f2() compared to f1().

The effect of this shift is that the resulting sound is delayed by 1/8192th of
a host sample. At 44100Hz that's 3ns, that should be inaudible ;-) BTW check
this excellent video for an explanation of how a sampled audio signal can be
delayed by less than the duration of a single sample:
   https://xiph.org/video/vid2.shtml

This patch only shifts the sinc functions. The next patch will actually exploit
the symmetry. That makes this patch _very_ small. But the reason why I still
choose to split it is that this part results in a (very small) change in the
output signal, while all later patches in this series should leave the result
unchanged.