[1b/2] Halve hq-resampler table size
The previous patch slightly changed (shifted) the values stored in the 2D table
of shifted sinc functions. This patch exploits some symmetry properties so that
we only need to store half the amount of data.

The size of this table depends on the ratio between the input and output sample
frequency. For resampling PSG (with native frequency 224kHz) to 44.1kHz we need
a table that's over 3MB in size! To resample MSX-MUSIC (native frequency
49.7kHz) the table is slightly over 700kB. So cutting the size of these tables
in half is certainly worth it.

The commit message of the previous patch may have suggested that we only store
half of each row in the table (half of each sinc). But that's not the case (and
not possible because in general a sampled shifted sinc has no (sample-aligned)
symmetry axis). Instead we exploit the property that shifting over 'k' is the
same as shifting the left-right-flipped function over '-k'.

So instead of storing half of each row in the table we store half the amount of
rows. And to re-calculate those dropped rows we read the previous rows in
reverse (both from right-to-left and from bottom-to-top).

Our boosted machine configurations have many different soundchips with
different native sample frequencies (PSG, SCC, MSX-MUSIC, MoonSound-wave and
-FM part). On those machines this patch saves over 3.2MB memory. But also on
simple MSX1 machines this patch saves over 1.5MB.

In some cases this patch also increases execution speed. On a i7 with 8MB L3
cache I couldn't measure any significant speed difference, but on a i5 with
only 4MB L3 cache this patch resulted in a 4% total emulation time speedup (of
course depending on what other parts of the emulation (or even the whole
system) are competing for cache).