Tweak SDLRasterizer::translateX()
Optimize division (by 2 or 4) into right-shift (over 1 or 2). Also because
division has the wrong and shift has the right rounding behavior for negative
values (division rounds towards zero, shift always rounds down).

I spend time on this function because I mistakenly though this function was
much more expensive than it really is: in a profile the division instruction
was taking ~25% of the drawDisplay() function which was taking ~5% of total
emulation time. Though that 5% was including child-functions instead time spend
in the function itself. Nevertheless this change does result in a small speedup
and it also makes the code slightly simpler (IMHO).