Tweak KUC implementation
We received the KUC VHDL source code of Manuel Pazos (thanks!). Based on this
we could fix some details in the emulation. Some were bugs in the documentation
or things Manuel forgot about himself ;-)

This patch changes the savestate format for the KUC mapper. But because it's
relatively new (and not yet included in any release) I didn't bother adding
backwards-compatibility code for the old format.