Introduce extra command line parsing phase: PHASE_DEFAULT_MACHINE
A commandline like
  openmsx -machine Boosted_MSXturboR_with_IDE -hda ~/msx/my_hd.dsk
failed with
  Uncaught exception: No hard disk named 'hda'.
even though the specified machine does have an integrated IDE interface.

The problem was the order in which the parameters were parsed. Before this
patch '-machine' and '-hda' had the same priority, and because '-machine' was
specified before '-hda' it was processed first. Though that's the wrong order:
the '-hda' options must already be seen before the IDE interface gets
instantiated. And swapping the order on the command line indeed hides the
problem.

Solved by introducing a new parsing phase for the -machine option, that comes
after the (exiting) PHASE_BEFORE_MACHINE, but before loading of the default
machine.