Set default sound_driver on Windows to SDL.
We use a deprecated API for the DirectX sound driver on Windows and it has some
issues (repeating the buffer in an underrun). Also I've heard from several
people in the past that the SDL driver works fine. So we may remove the DirectX
sound_driver in the (near) future. In the mean time, let's not make it default.