Removed Thread::stop()
This was always a bad idea, since it can kill threads without proper
cleanup taking place. But we relied on it until now to stop the worker
threads used for blocking reads.

This was blocking migration to modern APIs like C++11 threading and
SDL2, which don't have an equivalent to SDL_KillThread().