Fix(?) compile error with libc++ (rather than libstdc++)
Std classes (like std::thread) cannot be forward declared in a portable way.
E.g. a vender is allowed to implement it as a template with extra template
parameters (with default values) or as an alias (typedef) to another class
(possibly to a class in a versioned namespace like libc++ seems to be doing).