[6/7] Directly use std::thread in CliConnection
There's still a problem with thread termination in CliConnection, more
specifically the stdio and socket versions. Before this patch, when using such
a connection and exiting openMSX you got a crash: actually std::terminate() was
called because the helper thread was not properly shut down. Now we do join()
the helper thread, but it doesn't properly stop yet (e.g. for stdio it only
stops when the next data is read from stdin).

This issue needs to be addressed, but that's something for another patch.