Fixed exit of stdio connection on Windows
Although we can't call Poller::poll() on Windows, we should still check
whether Poller::abort() was called or not.

Note that on Windows, an extra line of input is required to make the
worker thread exit. This is not ideal, but since earlier experiments
with stdio on Windows showed that it was not usable for controlling
openMSX from other processes, this is only a problem when using stdio
manually, which is a rare use case.