Eliminate some SDL_GetVideoSurface() calls
This is a (small) step towards SDL2 because SDL2 no longer offer this function
(this function doesn't make sense for SDL2 because there could be more than one
window).

I've implemented this on the main branch rather than the sdl2 branch because on
main the patch can be tested (ATM the sdl2 branch doesn't compile yet). And
even on it's own (independent from sdl1/2) this patch contains some nice
cleanups.

This patch removes SDL_GetVideoSurface() calls that are made to query the
resolution of the output window. They are replaced by
Display::getOutputScreenResolution(). There are other calls to
SDL_GetVideoSurface() to query the pixel format. Those should also be replaced,
but that's for a later patch.

More in detail, this patch:
- Adds Display::getOutputScreenResolution()
  and  Display::setOutputScreenResolution()
- Changes the member variable "VisibleSurface::renderSettings" to "display" (the
  render settings can be queried from Display).
- Moves the calculation of the full-screen-flag from {SDL,SDLGL}VisibleSurface
  to the common base class.
- Changes the member "variable OSDImageBasedWidget::gui" to
  "OSDWidget::display" (the OSDGUI can be queried from Display).
- Gives TouchPad an extra "display" member variable (and eliminates all SDL
  calls from the TouchPad code).
- Combines "DummyOutputRectange::getOutput{Width,Height}()" into a single
  "getOutputSize()" method. (More clear/compact code and as a bonus only 1
  instead of 2 virtual calls).