Changed setWindowTitle() into updateWindowTitle()
Now that we can query the window title, it makes more sense to send an
update notification without the title string. This would allow the
back-end to for example ask for a shorter or longer title depending on
the kind of window it is showing. There are no immediate plans for
this, but it shows the design is more flexible.