Fix #1042 openMSX crashes after loading starfighters LD
More in detail the combination resampler=hq with stereo sound output went
wrong.

There were actually two distinct problems introduced on 2016/08/27 in commits
  1321bd568  [1b/2] Halve hq-resampler table size
  1237d54eb  [2/2] Permute rows in table

The first had a wrong array-index scale factor in the SSE2 stereo HQ resample
code. This resulted in bad sound quality, but usually no crash.

The second introduced undefined behavior: we kept a pointer to an array on the
stack, later that array is moved but we keep using the (now invalid) pointer.
This is clearly a bug, but unfortunately in practice it often happened to work
fine.