SDL2: made {GL,SDL}Image code compile
In the future in SDLimage we should switch from SDL_Surface to SDL_Texture (if
we want to keep SDLImage at all). That should greatly simplify the code. But
for now I wanted to do the minimal thing to get the stuff compiling. And it
might even work, but untested for now.