Added text event handling to OSD console
I changed the argument of normalKey() to 32 bits, because not all
unicode chars fit in 16 bits.

I dropped the swallowing of unhandled non-meta keys by the console.
Since keys are not passed to the MSX anyway when the console is open,
the only thing this did was prevent them from triggering hotkeys and
I don't see why that would be good.