Fixed compilation on gcc-4.8
Problem was with the old C abs() function versus the C++ abs() function. The
former is defined in <stdlib.h> and can only handle int parameters. The latter
is defined in <cmath> and is templatized (can handle any numeric type). So
depending on which header you include abs() can or cannot handle floating
point. This gotcha will be fixed in C++17 (you'll always get the C++ version).
Apparently my compiler (gcc-6) already implemented that, so I didn't see this
problem.