Fix most "clang-tidy -check=modernize-pass-by-value" warnings
The idea here is that 'sink' parameters (those that will be copied by the
callee) should be passed by value rather than by const-reference. And then the
callee should move the parameter rather copy it.

Because the parameter is passed by value it's now possible the caller needs to
make a copy of it. In that case we trade one copy in the callee for one copy in
the caller and we gain nothing. But it's also possible the callee has an rvalue
that can be moved instead of copied. In that case a copy is replaced by a move.

This all assumes that moves are cheaper than copies. Very often this is indeed
the case. An exception is our TclObject class, there the copy and move
constructors are equally expensive (or equally cheap). So passing those by
const-reference is still better than passing by value, and I didn't follow the
clang-tidy suggestion.