Fixed worker thread join in CliConnection
It is possible for an instance of a CliConnection subclass to be
destructed without the start() method ever being called. In that case
the worker thread was never created and calling join() on the thread
object would throw system_error.

We now check whether the thread is joinable before calling join() on
it. The code that calls join() was moved from the subclasses to the
base class. This has several advantages:
- avoid duplicate code now that it's no longer a single statement
- more symmetry between start() and end()
- the 'thread' member is now private to the base class
- it avoids a double join() if SocketConnection::output() initiates
  a connection shutdown on I/O error