Added emulation of Tecall_MSX_RAMFILE_TM220
This is a cartridge with 16kB ROM (though 2nd half of the ROM is filled with
0xff) and 16kB RAM. The RAM is powered by a battery. It adds BASIC CALL
commands to load/save files to a small RAM disk. See manual for much more
details:

   http://www.msxarchive.nl/pub/msx/graphics/jpg/hardware/manuals/

The implementation of this cartridge is relatively straight-forward, EXCEPT for
the /WE signal of the RAM. This signal is computed from 3 output bits of a
shift-register. The data-input of this shift-register is a combination of 2
data-bits of the msx cartridge slot. And the clock-input of the shift-register
is calculated from the M1 and slot-select cartridge slot signals. Short summary
(see comments in code for details): the TM220 monitors which instructions are
fetched from ROM, only after a specific sequence it allows to write to RAM.

So to emulate this cartridge correctly it's important to have an accurate
emulation of the M1 signal. See previous patch for more details on that.

Thanks to 'Quibus' and 'l_oliveira' for measuring and reverse-engineering the
TM220 hardware schematic.

'Quibus' also implemented the initial emulation of this cartridge. I only
tweaked it a little and added the M1-write-enable stuff.